gofer: Support restore of deleted directories whose original path is occupied.
It is possible that the application can open an FD to a directory, delete the
directory, create another directory at that path and then checkpoint the
container. In such a scenario, our gofer restore logic will fail with EEXIST
because a directory already exists on the host at that location when we try to
recreate it.

Use a temporary filename to recreate the directory in such a scenario. This
directory is deleted again anyways.

This is similar to e3c4c4cad86f ("gofer: Support restore of deleted files whose
original path is occupied.")

In addition to this, this change also makes the following bug fixes:
- Adds a random suffix to the temporary filename used to recreate the file/dir.
  This helps if multiple (different) deleted open directories exist at the same
  position.
- Change dentry.name for the duration of the restoreDeleted*() function. Note
  that dentry.restoreFile() use dentry.name to re-open handles and control FD.
  Hence setting the correct dentry.name is crucial for the handles to point to
  the correct file.
- Added a syscall test, which when run in save/restore mode, tests these code
  paths for handling open deleted files/directories.

Updates #11425

PiperOrigin-RevId: 775406486