Fix directfs restore for deleted regular file when read handle is not available
In such a situation, we can not rely on traditional methods (like openHandle())
to open a readable handle. This is because, for directfs to open a handle it to
re-walk the file via the parent using openat(parentFD, name). This does not
work for deleted files, it will fail with ENOENT. runsc gofer works around this
by using /proc/self/fd/ to re-open the control FD in the desired mode. However,
the sentry does not have access to any procfs instance (for security). See
fcbc289a7ac1 ("runsc: umount /proc in the sandbox namespace").

This change makes directfs just use the control FD to fetch file data. The
control FD should not be used for IO. We make an exception here for S/R.

Fixes #11903

PiperOrigin-RevId: 782379453